renaming app to LNDR
