+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.friendindebt",
+    package = "com.lndr",
 )
 
 android_resource(
     name = "res",
-    package = "com.friendindebt",
+    package = "com.lndr",
     res = "src/main/res",
 )
