+++ android/app/build.gradle
@@ -98,7 +98,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.friendindebt"
+        applicationId "com.lndr"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
@@ -137,9 +137,9 @@ android {
 }
 
 dependencies {
-    compile project(':react-native-blur')
     compile project(':react-native-vector-icons')
     compile project(':react-native-randombytes')
+    compile project(':react-native-blur')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
