+++ android/app/src/main/java/com/lndr/MainApplication.java
@@ -1,10 +1,11 @@
-package com.friendindebt;
+package com.lndr;
 
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
 import com.oblador.vectoricons.VectorIconsPackage;
 import com.bitgo.randombytes.RandomBytesPackage;
+import com.cmcewen.blurview.BlurViewPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -26,7 +27,8 @@ public class MainApplication extends Application implements ReactApplication {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
             new VectorIconsPackage(),
-            new RandomBytesPackage()
+            new RandomBytesPackage(),
+            new BlurViewPackage()
       );
     }
