+++ android/app/src/main/res/values/strings.xml
@@ -1,4 +1,3 @@
 <resources>
-
-    <string name="app_name">FriendInDebt</string>
+    <string name="app_name">LNDR</string>
 </resources>
