+++ ios/LNDR/Images.xcassets/AppIcon.appiconset/Contents.json
