+++ ios/LNDRTests/LNDRTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface FriendInDebtTests : XCTestCase
+@interface LNDRTests : XCTestCase
 
 @end
 
-@implementation FriendInDebtTests
+@implementation LNDRTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
